1.1.1 Update
Changelog for 1.1.1:
= Fixed Bulk Biceps and Discord to meet the new standard Swap power
wording.
= All Goals that included "from your hand" in the text no longer do
= Goals no longer go on the bottom of the deck but to a separate discard
pile
= Updated Chancellor Puddinghead's Entourage, Queen Platinum's Court,
and Commander Hurricane's Army to also no longer say "from your hand"

Changelog for 1.0.5:
CARD RULING UPDATES
= For all Gender and Race Change ships, the word "counts as" has become
"becomes" to eliminate confusion as to whether it keeps its former
designation.
= Derpy Hooves has been fixed to refer to "Fanfic Author Twilight"
instead of "Fan Fic Author Twilight"
= Princess Pile has been reworded slightly to be more rules appropriate.

= All Swap powers no longer have "Powers do not activate." Instead,
powers not activating is the norm. Powers only activate if a card says
so.
= Reworded Replace powers to be a little clearer.
= Love Poisons have been slightly reworded to be clearer.
= Friendship is Benefits has been changed to a 3-4 point Goal, where you
earn 4 points if there are no duplicates in your chain.
= Wording on the Twilestia series and "Just Experimenting" and I Swear
I'm Not Gay have been made consistent for their requirements, with "any
____" instead of "any one ____" etc.

CONSISTENCY
= Punctuation on Goals has been made consistent. All Goal requirements
will end in a period unless the requirement ends in a symbol.
= All card titles and ficlet titles have been made consistently Title
Case

RULES CHANGES
= Winning points totals have increased 1 point all around, so that you
definitely have to win at least 3 cards to win

NAME CHANGES
= Major General Rainbow Dash has been renamed Major General Dash due to
eliminate the new line in the name

AESTHETIC CHANGES
= The art of Invasive Species has been updated to reflect the correct
Goal back color.
= Fonts have been updated to ones that are better designed and
customized thanks to Lauren Herda. The new fonts are a better
streamlined Barthowheel for the title, Cabin Medium for the body text,
and Klinic Slab Italic for the flavor text. Thank you, Lauren! = And
finally, the Goal symbol has been updated to be a light blue on the
check instead of a pinkish purple.